Finished code for doing multinomial relative connectivity model.
Also added a test for multinomial stuff.

A bunch of other small modications to .R files to meet CRAN's demanding
standards (particularly to replace T with TRUE in many cases).

Fixes to DESCRIPTION to suggest mcmc

Renaming gamma.mean.sd.shape.scale to gammaParamsConvert to avoid it
being treated as an S3 Method of gamma function.